+++ src/datastructures.jl
@@ -581,3 +581,293 @@ or in operational period `t`.
 EMB.deficit_penalty(n::MultipleBuildingTypes) = n.penalty_deficit
 EMB.deficit_penalty(n::MultipleBuildingTypes, p::Resource) = n.penalty_deficit[p]
 EMB.deficit_penalty(n::MultipleBuildingTypes, t, p::Resource) = n.penalty_deficit[p][t]
+
+"""
+    ResourceBio{T<:Real} <: Resource
+
+Resources that can be transported and converted.
+These resources **cannot** be included as resources that are emitted, *e.g*, in the variable
+[`emissions_strategic`](@ref man-opt_var-emissions). Compared to a `ResourceCarrier`, the
+`ResourceBio` `Resource` includes additionally the fuel definition (a string identifier of
+the biomass) and the moisture content of the biomass (as a mass fraction).
+
+# Fields
+- **`id`** is the name/identifyer of the resource.
+- **`bio_type::String`** is the type of biomass, *e.g.*, "spruce_stem", "spruce_bark",
+  "spruce_T&B", or "birch_stem".
+- **`moisture::Float64`** is the moisture content of the biomass resource as a mass fraction.
+- **`co2_int::T`** is the CO₂ intensity, *e.g.*, t/MWh.
+"""
+struct ResourceBio{T<:Real} <: Resource
+    id::Any
+    bio_type::String
+    moisture::Float64
+    co2_int::T
+end
+
+"""
+    bio_type(p::ResourceBio)
+
+Returns the biomass type of a [`ResourceBio`](@ref) `p`.
+"""
+bio_type(p::ResourceBio) = p.bio_type
+
+"""
+    bio_type(p::ResourceBio)
+
+Returns the moisture content of a [`ResourceBio`](@ref) `p`.
+"""
+moisture(p::ResourceBio) = p.moisture
+
+"""
+    BioCHP <: NetworkNode
+
+A [`BioCHP`](@ref) node that samples the CHP model at https://github.com/iDesignRES/CHP_modelling.git.
+The `BioCHP` utilizes a linear, time independent conversion rate of the `input`
+[`Resource`](@ref)s to the output [`Resource`](@ref)s, subject to the available capacity.
+The capacity is hereby normalized to a conversion value of 1 in the fields `input` and
+`output`.
+
+# Fields
+- **`id`** is the name/identifier of the node.
+- **`cap::TimeProfile`** is the installed capacity.
+- **`electricity_resource::Resource`** is the electric power resource.
+- **`opex_var::TimeProfile`** is the variable operating expense per per capacity usage
+  through the variable `:cap_use`.
+- **`opex_fixed::TimeProfile`** is the fixed operating expense per installed capacity
+  through the variable `:cap_inst`.
+- **`input::Dict{<:Resource,<:Real}`** are the input [`Resource`](@ref)s with conversion
+  value `Real`.
+- **`output::Dict{<:Resource,<:Real}`** are the generated [`Resource`](@ref)s with
+  conversion value `Real`.
+- **`data::Vector{Data}`** is the additional data (*e.g.*, for investments). The field `data`
+  is conditional through usage of a constructor.
+"""
+struct BioCHP <: NetworkNode
+    id::Any
+    cap::TimeProfile
+    electricity_resource::Resource
+    opex_var::TimeProfile
+    opex_fixed::TimeProfile
+    input::Dict{<:ResourceBio,<:Real}
+    output::Dict{<:Resource,<:Real}
+    data::Vector{<:Data}
+end
+function BioCHP(
+    id,
+    cap::TimeProfile,
+    electricity_resource::Resource,
+    opex_var::TimeProfile,
+    opex_fixed::TimeProfile,
+    input::Dict{<:ResourceBio,<:Real},
+    output::Dict{<:Resource,<:Real},
+)
+    return BioCHP(
+        id,
+        cap,
+        electricity_resource,
+        opex_var,
+        opex_fixed,
+        input,
+        output,
+        [EmissionsEnergy()],
+    )
+end
+
+"""
+    BioCHP(
+        id::Any,
+        cap::TimeProfile,
+        mass_fractions::Dict{<:ResourceBio,<:Real},
+        heat_resources::Dict{<:ResourceHeat,<:Real},
+        electricity_resource::Resource;
+        data::Vector{Data} = Data[],
+        libpath::String = joinpath(
+            @__DIR__,
+            "..",
+            "..",
+            "CHP_modelling",
+            "build",
+            "lib",
+            "libbioCHP_wrapper.so",
+        ),
+    )
+
+Constructs a [`BioCHP`](@ref) instance where the power and heat production profiles are
+sampled from the `bioCHP_plant_c` function in the C++ library `CHP_modelling` with shared
+library file located at `libpath`. The BioCHP has electricity production of the resource
+`electricity_resource` and heat production of the resources in `heat_resources`
+(which can be different `ResourceHeat`s at different temperature levels).
+
+# Arguments
+- **`id`** is the name or identifier of the node.
+- **`cap`** is the installed electric capacity.
+- **`mass_fractions`** is the mass fractions of each input `ResourceBio`.
+- **`heat_resources`** is the output heat `ResourceHeat`s with the ratio of installed
+  capacity of heat to that of the electricity.
+- **`electricity_resource`** is the `Resource` for the electricity.
+
+# Keyword arguments
+- **`data::Vector{Data}`** is the additional data (*e.g.*, for investments). The field `data`
+  is conditional through usage of a constructor.
+- **`libpath`** is the absolute path of the `CHP_modelling` library file.
+
+!!! note "Current limitations"
+    - The moisture content of the biomass resource is currently not used in the model.
+    - The temperatures of `heat_resources` are currently not utilized as this would require
+      the specification of the individual heat production capacities.
+
+!!! note ""EmissionsEnergy"
+    If `EmissionsEnergy` is not included in the `data` field, it is automatically added.
+"""
+function BioCHP(
+    id::Any,
+    cap::TimeProfile,
+    mass_fractions::Dict{<:ResourceBio,<:Real},
+    heat_resources::Dict{<:ResourceHeat,<:Real},
+    electricity_resource::Resource;
+    data::Vector{Data} = Data[],
+    libpath::String = joinpath(
+        @__DIR__,
+        "..",
+        "..",
+        "CHP_modelling",
+        "build",
+        "lib",
+        "libbioCHP_wrapper.so",
+    ),
+)
+
+    # Get the capacity
+    el_capacity = cap.val
+
+    # fuel_def: name of each biomass feedstock
+    fuel_def_strings = [bio_type(res) for res ∈ keys(mass_fractions)]
+
+    # Create pointers
+    fuel_def_buffers = [Vector{UInt8}(string(s, '\0')) for s ∈ fuel_def_strings]
+    fuel_def_ptrs = [pointer(buf) for buf ∈ fuel_def_buffers]
+    fuel_def_ptr_array = pointer(fuel_def_ptrs)
+
+    # Create mass fractions from input that sum up to 1
+    resources = keys(mass_fractions)
+    normalization = sum(mass_fractions[res] for res ∈ resources)
+    normalized_mass_fractions = Dict{ResourceBio,Float64}(
+        res => val / normalization for (res, val) ∈ mass_fractions
+    )
+
+    # Yj: mass fraction of each biomass feedstock
+    # W_el: electric power output (MW_el)
+    # Qk: heat demand (MW)
+    # Tk_in: Return temperature for each heat demand (district heating)
+    # Tk_in: Supply temperature for each heat demand (district heating)
+    Qk_dict::Dict{Resource,Real} = Dict{Resource,Real}(
+        res => val * el_capacity for (res, val) ∈ heat_resources
+    )
+    Yj::Vector{Cdouble} = [normalized_mass_fractions[res] for res ∈ resources]
+    YH2Oj::Vector{Cdouble} = [moisture(res) for res ∈ resources]
+    W_el::Cdouble = el_capacity
+    Qk::Vector{Cdouble} = []
+    Tk_in::Vector{Cdouble} = []
+    Tk_out::Vector{Cdouble} = []
+    for (resource, val) ∈ Qk_dict
+        # Get the heat demand
+        push!(Qk, val)
+
+        # Get the supply temperature
+        supply_heat_profile = EMH.t_supply(resource)
+        if !isa(supply_heat_profile, FixedProfile)
+            @error "Current implementation require the supply heat profile to be fixed."
+        else
+            push!(Tk_in, supply_heat_profile.val)
+        end
+
+        # Get the return temperature
+        return_heat_profile = EMH.t_return(resource)
+        if !isa(return_heat_profile, FixedProfile)
+            @error "Current implementation require the return heat profile to be fixed."
+        else
+            push!(Tk_out, return_heat_profile.val)
+        end
+    end
+
+    # Preallocate output variables
+    # Mj: Required mass flow of each biomass feedstock
+    # C_inv: Investment cost
+    # C_op: Variable operating cost
+    Mj::Vector{Cdouble} = zeros(length(Yj))   # Output vector
+    Q_prod::Ref{Cdouble} = Ref{Cdouble}(0.0)  # Output double
+    W_el_prod::Ref{Cdouble} = Ref{Cdouble}(0.0)  # Output double
+    C_inv::Ref{Cdouble} = Ref{Cdouble}(0.0)  # Output double
+    C_op::Ref{Cdouble} = Ref{Cdouble}(0.0)   # Output double
+    C_op_var::Ref{Cdouble} = Ref{Cdouble}(0.0)   # Output double
+
+    # Get lengths of the vectors
+    len_Yj::Cint = length(Yj)
+    len_YH2Oj::Cint = length(YH2Oj)
+    len_Qk::Cint = length(Qk)
+    len_Tk_in::Cint = length(Tk_in)
+    len_Tk_out::Cint = length(Tk_out)
+    len_Mj::Cint = length(Mj)
+    len_fuel_def_ptrs::Cint = length(fuel_def_ptrs)
+
+    # Load the library if it's not already cached
+    if !haskey(LIB_CACHE, libpath)
+        @info "Loading the C module $libpath"
+        LIB_CACHE[libpath] = Libdl.dlopen(libpath)
+    end
+    lib = LIB_CACHE[libpath]
+    lib = Libdl.dlopen(libpath)
+
+    # Call the shared C function from the loaded library
+    @ccall $(@dlsym(lib, :bioCHP_plant_c))(
+        fuel_def_ptr_array::Ptr{Cstring}, len_fuel_def_ptrs::Cint,
+        Yj::Ptr{Cdouble}, len_Yj::Cint,
+        YH2Oj::Ptr{Cdouble}, len_YH2Oj::Cint,
+        W_el::Cdouble,
+        Qk::Ptr{Cdouble}, len_Qk::Cint,
+        Tk_in::Ptr{Cdouble}, len_Tk_in::Cint,
+        Tk_out::Ptr{Cdouble}, len_Tk_out::Cint,
+        Mj::Ptr{Cdouble}, len_Mj::Cint,
+        Q_prod::Ref{Cdouble},
+        W_el_prod::Ref{Cdouble},
+        C_inv::Ref{Cdouble},
+        C_op::Ref{Cdouble},
+        C_op_var::Ref{Cdouble},
+    )::Bool
+
+    input_updated = Dict{ResourceBio,Real}(
+        res => Mj[i] / W_el_prod[] for (i, res) ∈ enumerate(resources)
+    )
+    cap_updated = FixedProfile(Float64(W_el_prod[]))
+
+    opex_fixed = FixedProfile(Float64(C_op[]))
+    opex_var = FixedProfile(Float64(C_op_var[] / 8760))
+
+    output = Dict{Resource,Real}(
+        resource => val / W_el_prod[] for (resource, val) ∈ Qk_dict
+    )
+    output[electricity_resource] = 1.0
+
+    if !(EmissionsEnergy ∈ typeof.(data))
+        push!(data, EmissionsEnergy())
+    end
+
+    return BioCHP(
+        id,
+        cap_updated,
+        electricity_resource,
+        opex_var,
+        opex_fixed,
+        input_updated,
+        output,
+        data,
+    )
+end
+
+"""
+    electricity_resource(n::BioCHP)
+
+Returns the electricity resource of [`BioCHP`](@ref) node `n`.
+"""
+electricity_resource(n::BioCHP) = n.electricity_resource